+++ packages/toolkit/CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log - @contentauth/toolkit
 
-This log was last generated on Wed, 16 Apr 2025 17:11:17 GMT and should not be manually modified.
+This log was last generated on Tue, 06 May 2025 21:10:16 GMT and should not be manually modified.
+
+## 0.13.8
+Tue, 06 May 2025 21:10:16 GMT
+
+### Patches
+
+- Make templates optional in actions v2
+
+## 0.13.7
+Tue, 29 Apr 2025 23:13:41 GMT
+
+### Patches
+
+- Pass validation results through
+
+## 0.13.6
+Thu, 24 Apr 2025 22:18:28 GMT
+
+### Patches
+
+- Update c2pa rs
 
 ## 0.13.5
 Wed, 16 Apr 2025 17:11:17 GMT